src/requirements.txt
absl-py
airfrans
conda-pack
GitPython
h5py
keras
matplotlib
natsort
networkx
numpy
pandas
six
statsmodels
tensorboard
vtk
setuptools
pyvista
trimesh
circle_fit
torchtyping
wandb
timm
# open3d
rtree
einops
seaborn